refactor: assume activateExtension is running in Web Worker
Communicates using the Web Worker's global scope postMessage/MessageEvents by default.

BREAKING CHANGE: Extensions must omit the `MessageTransports` argument to `activateExtension`.
