+++ src/index.ts
@@ -1,3 +1,3 @@
-export { activateExtension } from './extension/extension'
+export { activateExtension } from './extension/extensionHost'
 export { SourcegraphExtensionAPI } from './extension/api'
 export * from './protocol'
